Add explanatory text
